Allow any kind of media in a single revision

+++ src/components/mixins/annotation_mixin.js
@@ -35,6 +35,14 @@ export const annotationMixin = {
       this.$options.undoneActionStack = []
     },
 
+    isEmptyCanvas () {
+      if (this.fabricCanvas) {
+        return this.fabricCanvas.getObjects().length > 0
+      } else {
+        return true
+      }
+    },
+
     getNewAnnotations (currentTime, annotation) {
       this.fabricCanvas.getObjects().forEach((obj) => {
         if (obj.type === 'path') {
@@ -120,7 +128,7 @@ export const annotationMixin = {
 
     undoLastAction () {
       const action = this.$options.doneActionStack.pop()
-      if (action) {
+      if (action && action.obj) {
         if (action.type === 'add') {
           this.deleteObject(action.obj)
         } else if (action.type === 'remove') {
@@ -244,6 +252,42 @@ export const annotationMixin = {
       if (this.fabricCanvas) {
         this.fabricCanvas.clear()
       }
+    },
+
+    addTypeArea () {
+      /** @lends fabric.IText.prototype */
+      // fix for : IText not editable when canvas is in a fullscreen
+      // element on chrome
+      // https://github.com/fabricjs/fabric.js/issues/5126
+      const originalInitHiddenTextarea =
+        fabric.IText.prototype.initHiddenTextarea
+      fabric.util.object.extend(fabric.IText.prototype, {
+        initHiddenTextarea: function () {
+          originalInitHiddenTextarea.call(this)
+          this.canvas.wrapperEl.appendChild(this.hiddenTextarea)
+        }
+      })
+    },
+
+    isFullScreen () {
+      return !!(
+        document.fullScreen ||
+        document.webkitIsFullScreen ||
+        document.mozFullScreen ||
+        document.msFullscreenElement ||
+        document.fullscreenElement
+      )
+    },
+
+    removeTypeArea () {
+      const originalInitHiddenTextarea =
+        fabric.IText.prototype.initHiddenTextarea
+      fabric.util.object.extend(fabric.IText.prototype, {
+        initHiddenTextarea: function () {
+          originalInitHiddenTextarea.call(this)
+          fabric.document.body.appendChild(this.hiddenTextarea)
+        }
+      })
     }
   }
 }
